+++ packages/terra-overlay/src/Overlay.jsx
@@ -56,6 +56,9 @@ const KEYCODES = {
   ESCAPE: 27,
 };
 
+
+let overlayClassName = '';
+
 const defaultProps = {
   children: null,
   isOpen: false,
@@ -109,6 +112,9 @@ class Overlay extends React.Component {
       for (let i = 0; i < this.containerChildren.length; i += 1) {
         prevTabIndex.push(this.containerChildren[i].tabIndex);
         this.containerChildren[i].tabIndex = -1;
+        if (this.containerChildren[i].className !== overlayClassName) { // childern[0] will be the overlay and it's content which should not be disabled for screen readers.
+          this.containerChildren[i].setAttribute('aria-hidden', 'true'); // prevent screen reader from moving to content behind the overlay
+        }
       }
       this.containerChildrenPrevTabIndex = prevTabIndex;
     }
@@ -118,6 +124,9 @@ class Overlay extends React.Component {
     if (this.containerChildren) {
       for (let i = 0; i < this.containerChildren.length; i += 1) {
         this.containerChildren[i].tabIndex = this.containerChildrenPrevTabIndex[i];
+        if (this.containerChildren[i].className !== overlayClassName) {
+          this.containerChildren[i].setAttribute('aria-hidden', 'false'); // Enables content on close of overlay
+        }
       }
     }
   }
@@ -175,7 +184,7 @@ class Overlay extends React.Component {
       customProps.className,
       `layer-${zIndexLayer}`,
     ]);
-
+    overlayClassName = OverlayClassNames;
     /*
       tabIndex set to 0 allows screen readers like VoiceOver to read overlay content when its displayed.
       Key events are added on mount.